requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.0.6
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.13.0
tzdata==2024.1
urllib3==2.3.0
whitenoise==6.7.0
